updated buttons to use basic port manipulation

+++ include/buttons.h
@@ -1,9 +1,12 @@
 #ifndef BUTTONS_H
 #define BUTTONS_H
 
+#include <avr/io.h>
+#include <stdint.h>
+
 namespace BUTTONS
 {
-    void init(int funcPin, int togglePin);
+    void init(uint8_t funcPin, uint8_t togglePin);
     bool isfPinPressed();
     bool istPinPressed();
 }
